4 лаба по защиет информации

+++ security_info/src/diffie_hellman.cpp
@@ -1,12 +1,11 @@
 #include "../include/info_protection.hpp"
 
-long long
-diffie_hellman_key_exchange(long long p, long long g, struct keys* key)
+long int diffie_hellman_key_exchange(long int p, long int g, struct keys* key)
 {
     if (g >= p)
         return -1;
-    long long Ya = mod_pow(g, key->a, p);
-    long long Yb = mod_pow(g, key->b, p);
+    long int Ya = mod_pow(g, key->a, p);
+    long int Yb = mod_pow(g, key->b, p);
 
     key->a = mod_pow(Yb, key->a, p);
     key->b = mod_pow(Ya, key->b, p);
